fix: 批量在线查询 device_flag IN 参数改用 uint16 修复 SQL 语法错误 (#61)
- db_online.go queryOnlineDevices: deviceFlags []uint8 -> []uint16。Go 的 []uint8 即 []byte，
  dbr 当二进制字面量处理，device_flag in ? 无法展开成 IN 列表，MySQL 1064 恒失败、
  mute_of_app 每次 fail-open 不抑制，静音功能实际失效
- service.go GetDeviceOnlines: flags 构建改为 []uint16
- queryOnlineDevices: uid 按 1000 分批，对齐相邻 getActiveByUIDs 的分批不变量，避免超大 IN
- 新增 service_device_onlines_test.go: DB-backed 回归，真执行 SQL 断言 Web/PC 在线命中、
  APP/离线被过滤，守住 IN 参数类型（纯函数单测/build/vet 均无法发现该缺陷）

+++ modules/user/service.go
@@ -1298,9 +1298,10 @@ func (s *Service) GetDeviceOnlines(uids []string, deviceFlags []config.DeviceFla
 	if len(uids) == 0 || len(deviceFlags) == 0 {
 		return nil, nil
 	}
-	flags := make([]uint8, 0, len(deviceFlags))
+	// 用 uint16 而非 uint8：dbr 会把 []uint8([]byte) 当二进制字面量，IN 列表无法展开（详见 queryOnlineDevices）。
+	flags := make([]uint16, 0, len(deviceFlags))
 	for _, f := range deviceFlags {
-		flags = append(flags, f.Uint8())
+		flags = append(flags, uint16(f.Uint8()))
 	}
 	models, err := s.onlineDB.queryOnlineDevices(uids, flags)
 	if err != nil {
